Added support for rel-payment links
Using the latest version of the UrlbarButton module this extension now supports rel-payment links in addition to API-lookups for detecting whether a page is flattrable or not.

The benefits of this is that unsubmitted content is now detectable as well and that the extension works fine in privacy mode as the extension still detects the rel-payment links then, but of course don't do any API-requests.