Added support for rel-canonical links
If the page URL isn't flattrable and no rel-payment links has been found then as a last attempt the canonical URL specified in the rel-canonical link tag will now also be sent to the Flattr API to be checked whether it is flattrable or note.

This means that if you go to eg. a YouTube video you will see a Flattr button even when YouTube have appended some kind of query variable to the URL for tracking how you navigated to that movie or something like that.

Of course this new feature also respects privacy mode and no requests will be made to the API when the browser is in privacy mode and no results will be cached.