Bigger rewrite to get rid of event mayhem
Previously the showForPage module would trigger for four-five events - now it triggers for only three and those three have three distinct listener types. This together with some other changes results in a simpler more understandable code where the check for whether a page is flattrable or not isn't triggered from everywhere and where the showing and hiding of a URL button isn't made all over the place but in a more controlled manor.

Bonus feature: New cool trigger that listens to new link elements in a already loaded page.